#9999: Move pybind impl to cpp
